fixed a few problems with running dsbUpdateWidget and loadPrefs asynchronously and improved some code

+++ amplissimus/lib/prefs.dart
@@ -1,80 +1,81 @@
-import 'dart:io';
-
 import 'package:amplissimus/logging.dart';
 import 'package:amplissimus/values.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 SharedPreferences preferences;
+Map<String, dynamic> edits = {};
 
-void waitForPrefs() {
-  while(preferences == null) sleep(Duration(milliseconds: 10));
+void set(String key, dynamic value) {
+  if(preferences == null) edits[key] = value;
+  else if(value is int) preferences.setInt(key, value);
+  else if(value is bool) preferences.setBool(key, value);
+  else if(value is String) preferences.setString(key, value);
+  else if(value is List<String>) preferences.setStringList(key, value);
+  else if(value is double) preferences.setDouble(key, value);
+  else throw '${value.runtimeType} cannot be saved in the SharedPreferences.';
 }
 
 int get counter {
-  waitForPrefs();
+  if(preferences == null) throw '';
   int i = preferences.getInt('counter');
   if(i == null) i = 0;
   return i;
 }
 set counter(int i) {
-  waitForPrefs();
-  preferences.setInt('counter', i);
+  set('counter', i);
 }
 String get username {
-  waitForPrefs();
+  if(preferences == null) throw '';
   String s = preferences.getString('username_dsb');
   if(s == null) s = '';
   return s;
 }
 set username(String s) {
-  waitForPrefs();
-  preferences.setString('username_dsb', s);
+  set('username_dsb', s);
 }
 String get password {
-  waitForPrefs();
+  if(preferences == null) throw '';
   String s = preferences.getString('password_dsb');
   if(s == null) s = '';
   return s;
 }
 set password(String s) {
-  waitForPrefs();
-  preferences.setString('password_dsb', s);
+  set('password_dsb', s);
 }
 String get grade {
-  waitForPrefs();
+  if(preferences == null) throw '';
   String s = preferences.getString('grade');
   if(s == null) s = '';
   return s;
 }
 set grade(String s) {
-  waitForPrefs();
-  preferences.setString('grade', s);
+  set('grade', s);
 }
 String get char {
-  waitForPrefs();
+  if(preferences == null) throw '';
   String s = preferences.getString('char');
   if(s == null) s = '';
   return s;
 }
 set char(String s) {
-  waitForPrefs();
-  preferences.setString('char', s);
+  set('char', s);
 }
 set designMode(bool isDarkMode) {
-  waitForPrefs();
-  preferences.setBool('is_dark_mode', isDarkMode);
+  set('is_dark_mode', isDarkMode);
 }
 
-void loadPrefs() async {
-  preferences = await SharedPreferences.getInstance();
-  bool isDarkMode = preferences.getBool('is_dark_mode');
-  ampInfo(ctx: 'Prefs', message: 'recognized isDarkMode = $isDarkMode');
-  if(isDarkMode == null) return;
-  AmpColors.setMode(isDarkMode);
+void loadPrefs() {
+  (() async => preferences = await SharedPreferences.getInstance())().then((thereisnovalue) {
+    bool isDarkMode = preferences.getBool('is_dark_mode');
+    ampInfo(ctx: 'Prefs', message: 'recognized isDarkMode = $isDarkMode');
+    AmpColors.setMode(isDarkMode);
+    edits.forEach((key, value) => set(key, value));
+    edits.clear();
+  });
 }
 
 void clear() {
-  waitForPrefs();
+  if(preferences == null) throw '';
   preferences.clear();
   ampInfo(ctx: 'Prefs', message: 'Cleared SharedPreferences.');
 }
